refactor: move common errors to sdk
