+++ demo/midgard-node/src/transactions/utils.ts
@@ -10,7 +10,6 @@ import { Data, Effect, Schedule } from "effect";
 import * as BlocksDB from "../database/blocks.js";
 import { Database } from "@/services/database.js";
 import { ImmutableDB } from "@/database/index.js";
-import { GenericErrorFields } from "@/utils.js";
 import { DBSelectError } from "@/database/utils/common.js";
 import { UnknownException } from "effect/Cause";
 
@@ -189,19 +188,19 @@ export const outRefsAreEqual = (outRef0: OutRef, outRef1: OutRef): boolean => {
 };
 
 export class TxSignError extends Data.TaggedError("TxSignError")<
-  GenericErrorFields & {
+  SDK.Utils.GenericErrorFields & {
     readonly txHash?: string;
   }
 > {}
 
 export class TxSubmitError extends Data.TaggedError("TxSubmitError")<
-  GenericErrorFields & {
+  SDK.Utils.GenericErrorFields & {
     readonly txHash?: string;
   }
 > {}
 
 export class TxConfirmError extends Data.TaggedError("TxConfirmError")<
-  GenericErrorFields & {
+  SDK.Utils.GenericErrorFields & {
     readonly txHash?: string;
   }
 > {}
