+++ demo/midgard-node/src/workers/utils/commit-block-header.ts
@@ -1,11 +1,6 @@
 import { Effect } from "effect";
 import * as SDK from "@al-ft/midgard-sdk";
 import { CML, Data, coreToUtxo, utxoToCore } from "@lucid-evolution/lucid";
-import {
-  CborDeserializationError,
-  CborSerializationError,
-  CmlUnexpectedError,
-} from "@/utils.js";
 
 export type WorkerInput = {
   data: {
@@ -55,13 +50,13 @@ export const serializeStateQueueUTxO = (
   stateQueueUTxO: SDK.TxBuilder.StateQueue.StateQueueUTxO,
 ): Effect.Effect<
   SerializedStateQueueUTxO,
-  CmlUnexpectedError | CborSerializationError
+  SDK.Utils.CmlUnexpectedError | SDK.Utils.CborSerializationError
 > =>
   Effect.gen(function* () {
     const core = yield* Effect.try({
       try: () => utxoToCore(stateQueueUTxO.utxo),
       catch: (e) =>
-        new CmlUnexpectedError({
+        new SDK.Utils.CmlUnexpectedError({
           message: `Failed to serialize UTxO`,
           cause: e,
         }),
@@ -69,7 +64,7 @@ export const serializeStateQueueUTxO = (
     const datumCBOR = yield* Effect.try({
       try: () => Data.to(stateQueueUTxO.datum, SDK.TxBuilder.StateQueue.Datum),
       catch: (e) =>
-        new CborSerializationError({
+        new SDK.Utils.CborSerializationError({
           message: `Failed to serialize datum`,
           cause: e,
         }),
@@ -85,7 +80,7 @@ export const deserializeStateQueueUTxO = (
   stateQueueUTxO: SerializedStateQueueUTxO,
 ): Effect.Effect<
   SDK.TxBuilder.StateQueue.StateQueueUTxO,
-  CmlUnexpectedError | CborDeserializationError
+  SDK.Utils.CmlUnexpectedError | SDK.Utils.CborDeserializationError
 > =>
   Effect.gen(function* () {
     const u = yield* Effect.try({
@@ -94,7 +89,7 @@ export const deserializeStateQueueUTxO = (
           CML.TransactionUnspentOutput.from_cbor_hex(stateQueueUTxO.utxo),
         ),
       catch: (e) =>
-        new CmlUnexpectedError({
+        new SDK.Utils.CmlUnexpectedError({
           message: `Failed to convert UTxO to CML: ${e}`,
           cause: e,
         }),
@@ -103,7 +98,7 @@ export const deserializeStateQueueUTxO = (
       try: () =>
         Data.from(stateQueueUTxO.datum, SDK.TxBuilder.StateQueue.Datum),
       catch: (e) =>
-        new CborDeserializationError({
+        new SDK.Utils.CborDeserializationError({
           message: `Failed to deserialize datum: ${e}`,
           cause: e,
         }),
